Move particle filter into particle_functor
