+++ lib/python/picongpu/picmi/particle_functor/__init__.py
@@ -6,4 +6,5 @@ License: GPLv3+
 """
 
 from .particle_functor import ParticleFunctor as ParticleFunctor, Particle as Particle
+from .particle_filter import ParticleFilter as ParticleFilter, FilteredSpecies as FilteredSpecies
 from .unit_dimension import UnitDimension as UnitDimension
